+++ src/main/java/christmas/domain/order/Date.java
@@ -1,4 +1,4 @@
-package christmas.domain;
+package christmas.domain.order;
 
 public class Date {
     private static final int MINIMUM_DATE = 1;
@@ -6,16 +6,16 @@ public class Date {
     private static final String INVALID_DATE_MESSAGE = "유효하지 않은 날짜입니다. 다시 입력해 주세요.";
     private final int date;
 
-    public Date(int date) {
+    public Date(final int date) {
         validate(date);
         this.date = date;
     }
 
-    private void validate(int date) {
+    private void validate(final int date) {
         validateRange(date);
     }
 
-    private void validateRange(int date) {
+    private void validateRange(final int date) {
         if (date < MINIMUM_DATE || date > MAXIMUM_DATE) {
             throw new IllegalArgumentException(INVALID_DATE_MESSAGE);
         }
