+++ src/main/java/christmas/domain/order/Order.java
@@ -1,5 +1,8 @@
-package christmas.domain;
+package christmas.domain.order;
 
+import christmas.domain.Menu;
+import christmas.domain.MenuCategory;
+import java.time.LocalDate;
 import java.util.Map;
 
 public class Order {
@@ -7,10 +10,10 @@ public class Order {
     private static final String INVALID_ORDER_MESSAGE = "유효하지 않은 주문입니다. 다시 입력해 주세요.";
     private static final String TOTAL_QUANTITY_ERROR_MESSAGE = "최대 20개의 메뉴만 주문 가능합니다.";
     private static final String ONLY_DRINKS_ORDERED_ERROR_MESSAGE = "음료만 주문할 수 없습니다.";
-    private final Date date;
+    private final LocalDate date;
     private final Map<Menu, Integer> menus;
 
-    public Order(final Date date, final Map<Menu, Integer> menus) {
+    public Order(final LocalDate date, final Map<Menu, Integer> menus) {
         validateMenus(menus);
         this.date = date;
         this.menus = menus;
@@ -52,7 +55,7 @@ public class Order {
         return total;
     }
 
-    public Date getDate() {
+    public LocalDate getDate() {
         return date;
     }
